using react navigator
