+++ example/android/settings.gradle
@@ -1,4 +1,10 @@
 rootProject.name = 'example'
+include ':react-native-screens'
+project(':react-native-screens').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-screens/android')
+include ':react-native-gesture-handler'
+project(':react-native-gesture-handler').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-gesture-handler/android')
+include ':react-native-reanimated'
+project(':react-native-reanimated').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-reanimated/android')
 include ':react-native-beacon'
 project(':react-native-beacon').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-beacon/android')
